xfs_io: add copy_file_range command
Add support for the copy_file_range() syscall.  The xfs_io interface for
the syscall is based on sendfile -- it writes to the current file from a
referenced path or other open file number.

This patch short circuits autoconf detection of the libc wrapper and
manually calls the raw syscall.  This would all be removed once the
syscall is fully merged and supported.

Signed-off-by: Zach Brown <zab@redhat.com>
Signed-off-by: Chandan Rajendra <chandan@linux.vnet.ibm.com>