mdrestore: Replace metadump header pointer argument with generic pointer type
We will need two variants of read_header(), show_info() and restore() helper
functions to support two versions of metadump formats. To this end, A future
commit will introduce a vector of function pointers to work with the two
metadump formats. To have a common function signature for the function
pointers, this commit replaces the first argument of the previously listed
function pointers from "struct xfs_metablock *" with "void *".

Signed-off-by: Chandan Babu R <chandan.babu@oracle.com>